docs: make a conda binary available in the environment
According to the snakemake docs, a `conda` binary is needed in the
environment:
https://snakemake.readthedocs.io/en/stable/getting_started/installation.html

snakemake 8.26 deprecated the use of conda-frontend (we were using
mamba).

We need to update the docs and remove the reference to
mac-test-environment.yaml

We're updating snakemake to >8.26, and setting constrains >9.0
