+++ constraints.txt
@@ -1,1 +1,1 @@
-snakemake==8.25.1
+snakemake>=9.0
