+++ pytest.ini
@@ -1,5 +1,6 @@
 [pytest]
 markers =
     short: marks tests as short (deselect with '-m "not short"')
-    R: marks tests that need R (deselect with '-m "not R"')
     conda: marks tests that need conda (deselect with '-m "not conda"')
+    envmodules: marks tests that need envmodules (deselect with '-m "not envmodules"')
+    apptainer: marks tests that need apptainer (deselect with '-m "not apptainer"')
