+++ test-environment.yml
@@ -4,11 +4,9 @@ channels:
   - nodefaults
 dependencies:
   - conda-forge::python >=3.12
-  - conda-forge::mamba >=1.5.8
-  - conda-forge::lmod >= 8.7.25
   - conda-forge::pip >= 24.1.2
-  # datrie will not be needed for snakemake 9.x
-  - conda-forge::datrie >= 0.8.2 # dep for snakemake, workaround for https://github.com/astral-sh/uv/issues/7525
+  - conda-forge::lmod >= 8.7.25
+  - conda-forge::conda >= 24.9.1 # this is not a typo, we install conda with micromamba
   - pip:
-     - poetry
-     - "."
+      - poetry # TODO: install omnibenchmark from pypi after #99 is done
+      - "."
