+++ tests/software/test_apptainer_cli.py
