+++ tests/software/test_conda.py
@@ -1,6 +1,6 @@
 ## updated from Snakemake's test suite
 ##
-## https://raw.githubusercontent.com/snakemake/snakemake/main/tests/tests.py
+##https://raw.githubusercontent.com/snakemake/snakemake/main/tests/tests.py
 ## Derivative of (c) 2012-2022 Johannes Köster johannes.koester@uni-due.com
 ## From Koester's license above:
 ##
@@ -11,23 +11,15 @@
 # OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 # DEALINGS IN THE SOFTWARE.
 
-import sys
 import os.path as op
-from utils.run import run
 
-sys.path.insert(0, op.dirname(__file__))
+from .runner import run
 
 
-## this won't work on Mac
-# TODO(ben): add a pytest conditional skip
-def test_easybuild_sys_toolchain_build():
+def test_conda():
     run(
-        Snakefile=op.join("04_easybuild_build_envmodules", "Snakefile"),
-        produced=op.join("binutils-2.35.eb_ld.txt"),
-        expected=op.join(
-            "04_easybuild_build_envmodules",
-            "expected_results",
-            "binutils-2.35.eb_ld.txt",
-        ),
-        method="envmodules",
+        Snakefile=op.join("00_conda", "Snakefile"),
+        produced=op.join("test0.out"),
+        expected=op.join("00_conda", "expected_results", "test0.out"),
+        method="conda",
     )
