+++ tests/software/test_pin_conda.py
