--- tests/software/utils/export_modules.sh
@@ -1,1 +1,0 @@
-../../../omni/software/utils/export_modules.sh
